pyproject.toml
[tool.poetry]
name = "SCRIPBOZO"
version = "0.1.0"
description = "A Twitch chatbot based on GPT2 trained on MOONMOON's chat logs"
authors = ["Kogasa <kogasapls@gmail.com>"]
readme = "README.md"
packages = [{ include = "scripbozo", from = "src" }]

[tool.poetry.scripts]
run = "scripbozo.__main__:main"

[tool.poetry.dependencies]
python = "^3.11"
torch = { version = "*", source = "torch_rocm" }
twitchio = "^2.5"
transformers = "^4.24"
datasets = "^2.7"
aiolimiter = "^1.0"
ujson = "^5.5"
aioschedule = "^0.5"
iso8601 = "^2.0"
requests = "^2.28"
python-dotenv = "^1.0"
pylint = { version = "*", optional = true }
black = { version = "*", optional = true }
reorder-python-imports = { version = "*", optional = true }
pre-commit = { version = "*", optional = true }

[tool.poetry.dev-dependencies]
pylint = "*"
black = "*"
reorder-python-imports = "*"
pre-commit = "*"

[[tool.poetry.source]]
name = "torch_rocm"
url = "https://download.pytorch.org/whl/rocm5.4.2"
priority = "explicit"

[tool.poetry.extras]
dev = ["pylint", "black", "reorder-python-imports"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
